fix s3 key regex and refactor code
