+++ server/src/lib/utils/validator.ts
@@ -1,8 +1,17 @@
 import ajv from 'ajv';
 import { createGenerator, DEFAULT_CONFIG } from 'ts-json-schema-generator';
 import { errorParser } from './error-parser';
+import zlib from 'zlib'
+import split2 from 'split2'
+import path, { join } from 'path';
 
 const schemas: { [key: string]: object } = {};
+
+export const typesPath = join(
+  path.dirname(require.resolve('@quintype/migration-helpers')),
+  'lib',
+  'editor-types.d.ts'
+);
 
 export function generateJsonSchema(
   filePath: string,
@@ -38,4 +47,22 @@ export function validator(type: string, typesPath: string, data: any): any {
     return errorParser(error, data['external-id'], type);
   }
   return 'valid';
+}
+
+export function asyncReadStream(file: any, type: string) {
+  return new Promise((resolve, reject) => {
+    let result: any = []
+    file
+    .pipe(zlib.createGunzip())
+    .pipe(split2(/\r?\n+/,JSON.parse))
+    .on('data', (obj: any) => {
+      result.push(validator(type, typesPath, obj))
+    })
+    .on('end', () => {
+      resolve(result)
+    })
+    .on('error', (e: any) => {
+      reject(e)
+    })
+  })
 }
